requirements.txt
Flask==2.3.2
Flask-Session==0.4.0
Werkzeug==2.3.2
python-decouple==3.8